excluding lstdc++fs
It is unnecessary this library because it is installed as part of the standard package. I had a bad time trying to compile it as is in my system (ARM64). If it is necessary for other systems, or versions, please feel free to delete this commit. Thanks for your amazing work.